Fixed tests to match source
